Move out other errors, bump version
